refactor: move utility method to utils
